GUI and various improvements

+++ app.js
@@ -1,15 +1,13 @@
 import { Color } from "three";
-import { Dexie } from "dexie";
 import { IfcViewerAPI } from "web-ifc-viewer";
 import {
    createOrOpenDatabase,
    loadIfInDB,
-   loadSavedIfc,
    preprocessAndSaveIfc,
    removeDatabase,
 } from "./components/db-load";
 import { selectMaterial } from "./components/three-utils";
-// import { getPropertySets } from "./components/get-properties";
+import { decodeIFCString } from "./components/utils";
 
 const container = document.getElementById("viewer-container");
 const viewer = new IfcViewerAPI({
@@ -17,20 +15,28 @@ const viewer = new IfcViewerAPI({
    backgroundColor: new Color(0xeeeeee),
 });
 
-//Create grid an axes
 viewer.grid.setGrid();
 viewer.axes.setAxes();
-
 selectMaterial(viewer);
 
 const previousData = localStorage.getItem("modelsNames");
-
 const db = createOrOpenDatabase();
-
 let properties;
 
 const loadButton = document.getElementById("load-button");
 const input = document.getElementById("file-input");
+setLoadButtonImg();
+function setLoadButtonImg() {
+   if (previousData) {
+      const path = document.getElementById("load-button-path");
+      let d = path.getAttribute("d");
+      path.setAttribute(
+         "d",
+         "M24 3.752l-4.423-3.752-7.771 9.039-7.647-9.008-4.159 4.278c2.285 2.885 5.284 5.903 8.362 8.708l-8.165 9.447 1.343 1.487c1.978-1.335 5.981-4.373 10.205-7.958 4.304 3.67 8.306 6.663 10.229 8.006l1.449-1.278-8.254-9.724c3.287-2.973 6.584-6.354 8.831-9.245z"
+      );
+   }
+}
+
 loadButton.onclick = () => {
    if (previousData) {
       removeDatabase(db);
@@ -50,17 +56,31 @@ async function loadPropertiesFromDB() {
    properties = JSON.parse(data);
 }
 
-// Get properties of selected item
+// Get properties of selected item and create menu
 window.onmousemove = () => viewer.IFC.selector.prePickIfcItem();
 window.addEventListener("dblclick", async () => {
+   if (dimensionsActive || clippingPlanesActive) {
+      return;
+   }
+   if (!propertiesMenuActive) {
+      propertiesButton.click()
+   }
    const result = await viewer.IFC.selector.pickIfcItem();
    if (result) {
       const foundProperties = properties[result.id];
       getPropertySets(foundProperties);
       createPropertiesMenu(foundProperties);
+      if (simpleChildSelected.li) {
+         simpleChildSelected.li.classList.remove("selected");
+      }
    } else {
       viewer.IFC.selector.unpickIfcItems();
+      removeAllChildren(propsGUI)
    }
+});
+window.addEventListener("keydown", (event) => {
+   if (event.code === "Escape") viewer.IFC.selector.unpickIfcItems();
+   removeAllChildren(propsGUI)
 });
 function getPropertySets(props) {
    const id = props.expressID;
@@ -80,7 +100,17 @@ function getPropertySets(props) {
    props.psets = psets;
 }
 
+const propertiesMenu = document.getElementById("ifc-property-menu");
+const treeMenu = document.getElementById("ifc-tree-menu");
+
+let propertiesMenuActive = false
 const propsGUI = document.getElementById("ifc-property-menu-root");
+const propertiesButton = document.getElementById("properties-button");
+propertiesButton.onclick = () => {
+   propertiesMenuActive = !propertiesMenuActive
+   propertiesButton.classList.toggle("button-active");
+   propertiesMenu.classList.toggle("hidden");
+};
 
 function createPropertiesMenu(properties) {
    removeAllChildren(propsGUI);
@@ -92,6 +122,7 @@ function createPropertiesMenu(properties) {
    for (let key in properties) {
       createPropertyEntry(key, properties[key]);
    }
+
 }
 
 function createPropertyEntry(key, value) {
@@ -107,7 +138,8 @@ function createPropertyEntry(key, value) {
 
    const valueElement = document.createElement("div");
    valueElement.classList.add("ifc-property-value");
-   valueElement.textContent = value;
+   const decodedValue = decodeIFCString(value);
+   valueElement.textContent = decodedValue;
    propContainer.appendChild(valueElement);
 
    propsGUI.appendChild(propContainer);
@@ -120,17 +152,18 @@ function removeAllChildren(element) {
 }
 
 //Spatial tree
-const toggler = document.getElementsByClassName("caret");
+const treeButton = document.getElementById("tree-button");
+treeButton.onclick = () => {
+   treeButton.classList.toggle("button-active");
+   treeMenu.classList.toggle("hidden");
+   constructTreeAndMenu();
+};
 
-for (let i = 0; i < toggler.length; i++) {
-   toggler[i].addEventListener("click", () => {
-      toggler[i].parentElement
-         .querySelector(".nested")
-         .classList.toggle("active");
-      toggler[i].classList.toggle("caret-down");
-   });
+async function constructTreeAndMenu() {
+   const SpatialTree = await constructSpatialTree();
+   createTreeMenu(SpatialTree);
 }
-////////////////////////////////
+
 // Utils functions
 function getFirstItemOfType(type) {
    return Object.values(properties).find((item) => item.type === type);
@@ -157,7 +190,6 @@ async function constructSpatialTree() {
 
    return ifcProjectNode;
 }
-
 // Recursively constructs the spatial tree
 async function constructSpatialTreeNode(item, contains, spatials) {
    const spatialRels = spatials.filter(
@@ -191,14 +223,18 @@ async function constructSpatialTreeNode(item, contains, spatials) {
 
    item.children = children;
 }
-///////////////
-const leftMenu = document.getElementById("menu-left");
-const propertiesButton = document.getElementById("properties-button");
-propertiesButton.onclick = () => {
-   propertiesButton.classList.toggle("button-active");
-   leftMenu.classList.toggle("hidden");
-   createTreeMenu(constructSpatialTree());
-};
+
+//Treemenu
+const toggler = document.getElementsByClassName("caret");
+
+for (let i = 0; i < toggler.length; i++) {
+   toggler[i].addEventListener("click", () => {
+      toggler[i].parentElement
+         .querySelector(".nested")
+         .classList.toggle("active");
+      toggler[i].classList.toggle("caret-down");
+   });
+}
 
 function createTreeMenu(ifcProject) {
    const root = document.getElementById("tree-root");
@@ -260,14 +296,16 @@ function createSimpleChild(parent, node) {
       viewer.IFC.selector.prepickIfcItemsByID(0, [node.expressID]);
    };
    childNode.onclick = async () => {
-      viewer.IFC.selector.pickIfcItemsByID(0, [node.expressID], true, true);
-      const props = await viewer.IFC.getProperties(
+      await viewer.IFC.selector.pickIfcItemsByID(
          0,
-         node.expressID,
+         [node.expressID],
          true,
-         false
+         true
       );
+      const props = properties[node.expressID];
+      getPropertySets(props);
       createPropertiesMenu(props);
+
       if (simpleChildSelected.li) {
          simpleChildSelected.li.classList.remove("selected");
       }
@@ -276,33 +314,23 @@ function createSimpleChild(parent, node) {
    };
 }
 
-function DecodeIFCString(ifcString) {
-   const ifcUnicodeRegEx = /\\X2\\(.*?)\\X0\\/giu;
-   let resultString = ifcString;
-   let match = ifcUnicodeRegEx.exec(ifcString);
-   while (match) {
-      const unicodeChar = String.fromCharCode(parseInt(match[1], 16));
-      resultString = resultString.replace(match[0], unicodeChar);
-      match = ifcUnicodeRegEx.exec(ifcString);
-   }
-   return resultString;
-}
-
 //Clipping planes
 let clippingPlanesActive = false;
 
 const clipperButton = document.getElementById("clip-button");
 
 clipperButton.onclick = () => {
+   if (dimensionsActive) {
+      measureButton.click();
+   }
    clippingPlanesActive = !clippingPlanesActive;
    viewer.clipper.active = clippingPlanesActive;
    clipperButton.classList.toggle("button-active");
 };
 
 window.addEventListener("dblclick", () => {
-   if(clippingPlanesActive)
-   viewer.clipper.createPlane()
-   viewer.context.renderer.postProduction.update()
+   if (clippingPlanesActive) viewer.clipper.createPlane();
+   viewer.context.renderer.postProduction.update();
 });
 
 window.addEventListener("keydown", (event) => {
@@ -318,22 +346,35 @@ let dimensionsPreviewActive = false;
 const measureButton = document.getElementById("measure-button");
 
 measureButton.onclick = () => {
+   if (clippingPlanesActive) {
+      clipperButton.click();
+   }
    dimensionsActive = !dimensionsActive;
    dimensionsPreviewActive = !dimensionsPreviewActive;
 
    viewer.dimensions.active = dimensionsActive;
    viewer.dimensions.previewActive = dimensionsPreviewActive;
-   measureButton.classList.toggle("button-active")
+   measureButton.classList.toggle("button-active");
 };
 
 window.addEventListener("dblclick", () => {
-   if(dimensionsActive)
-   viewer.dimensions.create()
+   if (dimensionsActive) viewer.dimensions.create();
 });
 
 window.addEventListener("keydown", (event) => {
-   if(event.code === 'Delete' && dimensionsActive)
-   viewer.dimensions.delete()
-   viewer.context.renderer.postProduction.update()
+   if (event.code === "Delete" && dimensionsActive) viewer.dimensions.delete();
+   viewer.context.renderer.postProduction.update();
 });
 
+//Settings screen
+const settingsButton = document.getElementById("settings-button");
+const settingsScreen = document.getElementById("settings-screen");
+
+settingsButton.onclick = () => {
+   settingsButton.classList.toggle("button-active");
+   settingsScreen.classList.toggle("hidden");
+};
+settingsScreen.onclick = () => {
+   settingsButton.classList.toggle("button-active");
+   settingsScreen.classList.toggle("hidden");
+};
